refactor test to use Security

+++ src/main/java/org/zeros/farm_manager_server/Domain/DTO/Crop/CropParameters/RapeSeedParametersDTO.java
@@ -3,6 +3,8 @@ package org.zeros.farm_manager_server.Domain.DTO.Crop.CropParameters;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @AllArgsConstructor
@@ -10,8 +12,8 @@ import lombok.experimental.SuperBuilder;
 @EqualsAndHashCode(callSuper = true)
 @SuperBuilder
 public class RapeSeedParametersDTO extends CropParametersDTO {
-    private float density;
-    private float humidity;
-    private float oilContent;
+    private BigDecimal density;
+    private BigDecimal humidity;
+    private BigDecimal oilContent;
 
 }
