+++ src/main/java/org/zeros/farm_manager_server/Domain/DTO/Crop/CropSaleDTO.java
@@ -5,6 +5,7 @@ import lombok.experimental.SuperBuilder;
 import org.zeros.farm_manager_server.Domain.DTO.BaseEntityDTO;
 import org.zeros.farm_manager_server.Domain.Enum.ResourceType;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -19,8 +20,8 @@ public class CropSaleDTO extends BaseEntityDTO {
     private UUID crop;
     private LocalDate dateSold;
     private String soldTo;
-    private float amountSold;
-    private float pricePerUnit;
+    private BigDecimal amountSold;
+    private BigDecimal pricePerUnit;
     private ResourceType resourceType;
     private String unit;
     private UUID cropParameters;
