+++ src/main/java/org/zeros/farm_manager_server/Domain/DTO/Data/SubsideDTO.java
@@ -4,6 +4,7 @@ import lombok.*;
 import lombok.experimental.SuperBuilder;
 import org.zeros.farm_manager_server.Domain.DTO.BaseEntityDTO;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Set;
 import java.util.UUID;
@@ -20,7 +21,7 @@ public class SubsideDTO extends BaseEntityDTO {
     private String description;
     private LocalDate yearOfSubside;
     private Set<UUID> speciesAllowed;
-    private float subsideValuePerAreaUnit;
+    private BigDecimal subsideValuePerAreaUnit;
 
     public void setYearOfSubside(LocalDate date) {
         yearOfSubside = LocalDate.ofYearDay(date.getYear(), 1);
