+++ src/main/java/org/zeros/farm_manager_server/Domain/DTO/Fields/FieldDTO.java
@@ -4,6 +4,7 @@ import lombok.*;
 import lombok.experimental.SuperBuilder;
 import org.zeros.farm_manager_server.Domain.DTO.BaseEntityDTO;
 
+import java.math.BigDecimal;
 import java.util.Set;
 import java.util.UUID;
 
@@ -16,13 +17,13 @@ import java.util.UUID;
 public class FieldDTO extends BaseEntityDTO {
     private String fieldName;
     private String description;
-    private float area;
+    private BigDecimal area;
     private UUID user;
     private UUID fieldGroup;
     private Set<UUID> fieldParts;
     private Set<String> surveyingPlots;
     private Boolean isOwnField;
-    private float propertyTax;
-    private float rent;
+    private BigDecimal propertyTax;
+    private BigDecimal rent;
     private Boolean isArchived;
 }
